refactor: major codebase improvements and UI fixes
- Refactored TransactionTriggers component into modular sub-components:
  - KlvTransfer, KdaTransfer, SmartContractInteraction, TransactionHistory
  - Created reusable TransactionResult component for consistent result display
  - Improved code organization and maintainability

- Fixed UI/UX issues:
  - Git Version Control section now uses tabbed interface to prevent overflow
  - Fixed Transaction Playground title appearing under header
  - Resolved duplicate success toast notifications
  - Fixed transaction status updates in history list

- Code quality improvements:
  - Centralized network configuration in constants/network.ts
  - Removed unused imports, functions, and assets
  - Fixed TypeScript type safety issues
  - Removed console.log statements from production code

- Enhanced transaction status tracking:
  - Added proper status update callbacks between components
  - Implemented mounted state checks to prevent memory leaks
  - Maintained transaction history persistence in localStorage

Breaking changes: None - all existing functionality preserved
